+++ src/routes/+error.svelte
@@ -1,22 +1,5 @@
-<script context="module">
-    export function load({ error, status }) {
-        const { message } = error
-        return {
-            props: {
-                title: status,
-                message,
-            },
-        }
-    }
-</script>
-
 <script>
-    export let title
-    export let message
+    import { page } from '$app/stores'
 </script>
 
-<section>
-    <h1>{title + ' Error'}</h1>
-
-    <p>{message}</p>
-</section>
+<h1>{$page.status} : {$page.error.message}</h1>
