+++ src/routes/events/+page.svelte
@@ -1,24 +1,10 @@
-<script context="module">
-    export async function load({ fetch }) {
-        try {
-            const events = await fetch('/events.json').then((r) => r.json())
-            return {
-                props: {
-                    events,
-                },
-            }
-        } catch (error) {
-            console.error(error)
-        }
-    }
-</script>
-
 <script>
     import Event from '$lib/components/images/Event.svelte'
     import Button from '$lib/components/buttons/Button.svelte'
     import SEO from '$lib/components/base/SEO.svelte'
 
-    export let events
+    export let data
+    const { events } = data
 </script>
 
 <SEO title="Events" />
