+++ src/routes/events/[slug]/+page.js
@@ -1,0 +1,5 @@
+export async function load({ params, fetch }) {
+    const { slug } = params
+    const event = await fetch(`/events/${slug}.json`).then((r) => r.json())
+    return { event }
+}
