+++ src/routes/journal/+page.js
@@ -1,0 +1,7 @@
+import { posts } from '$lib/services/posts'
+
+export async function load() {
+    return {
+        posts,
+    }
+}
