+++ src/routes/journal/+page.svelte
@@ -1,15 +1,3 @@
-<script context="module">
-    import { posts } from '$lib/services/posts'
-
-    export async function load() {
-        return {
-            props: {
-                posts,
-            },
-        }
-    }
-</script>
-
 <script>
     import JournalEntrySet from '$lib/components/journal/EntrySet.svelte'
     import SEO from '$lib/components/base/SEO.svelte'
