+++ src/routes/photography/+page.js
@@ -1,0 +1,7 @@
+export async function load({ fetch }) {
+    const { images } = await fetch('/photography.json').then((r) => r.json())
+
+    return {
+        images,
+    }
+}
